Cloud of war movement display
Current problem is that the player position displays but the sliced rows and elements around the player do not.
Also need to reinstate fog of war to areas that the player can no longer see.

For clearing the fog in the visibile area it might be a good idea to create the tiles on the initial map and then transfer the positions and values to the visible board (since the board currently only knows its own x,y positioning of tiles)