Preparing to deploy to heroku
Fixed the board disappearing when going to the top or the left side of the screen.
Disabled the fix because it causes the screen to expand and contract (maybe take a look at yFrom...).
Need to add logic that stops from moving into empty tiles, maybe use those as the walls instead.